Breakout expression engine from qualityquery
